Rename records_mover.records.types
Renamed to records_mover.records.records_types in order to avoid an
issue in mypy where in local IDE verification mode, it doesn't like
modules named "types".
