+++ tests/component/records/test_validated_records_hints.py
@@ -1,5 +1,5 @@
 from typing_inspect import get_args
-from records_mover.records.types import DelimitedVariant
+from records_mover.records.records_types import DelimitedVariant
 from records_mover.records import DelimitedRecordsFormat
 import unittest
